+++ packages/core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @staffroom/core
 
+## 0.1.1
+
+No changes in this release.
+
 ## 0.1.0
 
 ### Minor Changes